Add initial implementation for Start command
Start writes out a new entry (with optional note) into the configured
log file. The start time is the current time in seconds since the UNIX
epoch. This is not configurable.

Note, decided against adding an explicit "End" entry when running Start.
I think we'll instead treat a Start as an implict End when parsing for
stats summary.