Add initial complete command
Add `complete` command for entering a logging a previously completed
task in the log. This "replaces" the stubbed out `amend` command because
it seemed too complicated as previously envisioned.

Note, the current iteration of this does not guarantee that the user keeps
the log "valid". For example, you could run Start and then run Complete with
a start time (calculated or otherwise) that occurs prior to the Start command.
This would result in an invalid log that wouldn't be valid on parse. Will
need to address this in a future commit.