Update `end` to check if task is active
It is now an error to END a task if there is no active task
in the log (i.e., the last element in the log is not a START).
This should ensure that the log remains "valid" so we can make
some simplifying assumptions when processing for summary generation.