Fix dropped group bug in group_tasks()
Fix two related bugs in group_tasks that resulted in chunks of the log
being dropped when tasks were grouped. The last date encountered when
"rolling over" from one day to another is now properly updated.
Additionally, the task that caused the rollover is now properly added to
the group.

Additional minor name reformatting for consistency.